move prepare_for_extend to BaseDraftWorker; drop EagleDraftExtendInputV2Mixin
