+++ python/sglang/srt/speculative/eagle_info_v2.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING
 
 import torch
 import torch.nn.functional as F
@@ -61,7 +61,6 @@ if TYPE_CHECKING:
         EAGLEDraftCudaGraphRunner,
     )
     from sglang.srt.speculative.eagle_info import (
-        EagleDraftExtendInput,
         EagleDraftInput,
         EagleVerifyInput,
     )
@@ -305,77 +304,6 @@ class EagleDraftInputV2Mixin:
         forward_batch = ForwardBatch.init_new(batch, draft_model_runner)
         can_cuda_graph = cuda_graph_runner and cuda_graph_runner.can_run(forward_batch)
         return forward_batch, can_cuda_graph
-
-
-class EagleDraftExtendInputV2Mixin:
-    def prepare_for_extend_to_fill_draft_kvcache(
-        self: EagleDraftExtendInput,
-        batch: ScheduleBatch,
-        predict: torch.Tensor,
-        num_draft_tokens: int,
-        draft_model_runner: Any,
-        cuda_graph_runner: Any,
-    ):
-        bs = len(batch.seq_lens)
-        extend_num_tokens = bs * num_draft_tokens
-        # When seq_lens_cpu is absent, stay on GPU-only path -- no .tolist()/.cpu().
-        gpu_only = batch.seq_lens_cpu is None
-
-        batch.spec_info = self
-        batch.input_ids = predict
-        maybe_detect_oob(
-            batch.input_ids,
-            0,
-            batch.model_config.vocab_size,
-            "v2 prepare_for_extend_to_fill_draft_kvcache input_ids",
-        )
-        # init_new requires both list or both Tensor;
-        # gpu_only emits device tensors to skip H2D.
-        if gpu_only:
-            batch.prefix_lens = batch.seq_lens.to(torch.int32)
-            batch.extend_lens = torch.full(
-                (bs,), num_draft_tokens, dtype=torch.int32, device=batch.seq_lens.device
-            )
-        else:
-            batch.prefix_lens = batch.seq_lens_cpu.tolist()
-            batch.extend_lens = [num_draft_tokens] * bs
-        batch.extend_num_tokens = extend_num_tokens
-        capture_mode = (
-            CaptureHiddenMode.NULL
-            if draft_model_runner.spec_algorithm.is_standalone()
-            else CaptureHiddenMode.FULL
-        )
-        batch.forward_mode = (
-            ForwardMode.IDLE
-            if batch.forward_mode.is_idle()
-            else ForwardMode.DRAFT_EXTEND_V2
-        )
-        batch.capture_hidden_mode = capture_mode
-        forward_batch = ForwardBatch.init_new(batch, draft_model_runner)
-        # Forward sees post-write length (draft extend writes num_draft_tokens
-        # slots); mutation stays on forward_batch to preserve SB.seq_lens.
-        forward_batch.seq_lens = forward_batch.seq_lens + num_draft_tokens
-        if not gpu_only:
-            forward_batch.seq_lens_cpu = forward_batch.seq_lens_cpu + num_draft_tokens
-            forward_batch.seq_lens_sum = int(forward_batch.seq_lens_cpu.sum())
-        else:
-            # Supply CPU mirror (extend_seq_lens are all num_draft_tokens) so
-            # backend max() reads from list without a per-iter D2H sync.
-            forward_batch.extend_seq_lens_cpu = [num_draft_tokens] * bs
-        can_cuda_graph = cuda_graph_runner and cuda_graph_runner.can_run(forward_batch)
-        if not batch.forward_mode.is_idle() and not can_cuda_graph:
-            draft_model_runner.attn_backend.init_forward_metadata(forward_batch)
-            # Planned pre-pad; do NOT opt into post-pad re-plan. DSA's indexer
-            # cannot rebuild its deep_gemm schedule_meta on a DP-padded batch
-            # (the `_batch_size == batch_size` assertion, see #27091); the
-            # marked pre-pad metadata is used as-is, matching the proven
-            # skip_attn_backend_init=True behavior.
-            # On NPU with --disable-cuda-graph, block_table shape won't match
-            # after prepare_mlp_sync_batch padding; defer re-init to
-            # forward_extend (post-pad) instead.
-            if not _is_npu or can_cuda_graph:
-                forward_batch.mark_forward_metadata_ready()
-        return forward_batch
 
 
 @dataclass
